Rewrite, refactor, mind the spec
This is a rewrite of most of the old code and a refactor of much of what
remained. Reason for the rewrite being the realization that TTML is not
just a loose collection of conventions but has an actual spec.

Still not a general purpose TTML parser though, and still most likely to
have serious blindspots.